nccl_ofi_tuner: Tuner component for aws-ofi-nccl
This commit introduces a custom tuner plugin for NCCL that bypasses the
cost function used by NCCL today. The existing cost function is a
combination of empirically derived costs, lat/bw reported by network
plugins, and other coefficients that are unexplainable. These defaults
seem to work well for IB, but not for EFA. The tuner is currently built
and used only when running on AWS.

In this initial implementation, a simple Hockney model is used as a
starting point, and only AllReduce is supported. Once we prove out the
viability of the tuner plugin in selecting algorithm switch points, we
can extend this to LogGP-based models to account for protocol specifics
used in the aws-ofi-nccl plugin (such as the eager->rendezvous switch
and the cost of control message exhange).

At the time of this commit, NCCL explicitly dlopen()s the
tuner plugin with a path provided to NCCL_TUNER_PLUGIN, but that will
change in the future to automatically pick up a tune from the network
plugin install path (v2.21 likely). NCCL should also look for the tuner
symbols with dlsym() implicitly if the environment variable is not set,
but that's a separate discussion to be had with the NCCL team.

Note that nccl-headers/tuner.h had to be slightly modified to removet
redundancy with definitions in nccl-headers/net.h and
nccl-headers/error.h. There are still some redundant definitions, but in
the interest of keeping the header from NCCL as close to the upstream
version as possible, this commit does not refactor the common
definitions for reuse.

Signed-off-by: Raghu Raja <raghunch@amazon.com>